Web App/Functions/classifier/requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.2.5
attr==0.3.1
attrs==19.3.0
azure-functions==1.2.1
boto3==1.10.2
botocore==1.13.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.20
dill==0.3.1.1
docutils==0.15.2
filelock==3.0.12
ftfy==5.7
future==0.18.1
gast==0.2.2
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.24.3
h5py==2.10.0
idna==2.8
imageio==2.5.0
isort==4.3.21
jmespath==0.9.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
nltk==3.4.5
numpy==1.17.1
opt-einsum==3.1.0
pandas==0.25.1
Pillow==6.1.0
promise==2.2.1
protobuf==3.10.0
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.35
scikit-learn==0.21.3
scipy==1.3.1
sentencepiece==0.1.83
six==1.12.0
sklearn==0.0
tensorboard==2.0.0
tensorflow==2.0.1
tensorflow-datasets==1.3.0
tensorflow-estimator==2.0.1
tensorflow-metadata==0.15.0
termcolor==1.1.0
tokenizers==0.5.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.6.0+cpu
tqdm==4.36.1
transformers==2.5.1
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
